Refactor project structure folders
