+++ BankApp/app/src/main/java/br/com/rphmelo/bankapp/common/utils/Variables.kt
@@ -1,0 +1,6 @@
+package br.com.rphmelo.bankapp.common.utils
+
+object Variables {
+    val ROOT_URL = "https://bank-app-test.herokuapp.com/api/"
+
+}
