+++ BankApp/app/src/main/java/br/com/rphmelo/bankapp/currency/domain/models/CurrencyView.kt
@@ -1,0 +1,11 @@
+package br.com.rphmelo.bankapp.currency.domain.models
+
+interface CurrencyView {
+    fun showProgress()
+    fun hideProgress()
+    fun setupToolbar()
+    fun setToolbarData()
+    fun setToolbarError()
+    fun setStatementListData()
+    fun setStatementListError()
+}
