+++ BankApp/app/src/main/java/br/com/rphmelo/bankapp/currency/domain/models/OnCurrencyLoadDataListener.kt
@@ -1,0 +1,9 @@
+package br.com.rphmelo.bankapp.currency.domain.models
+
+interface OnCurrencyLoadDataListener {
+    fun onSetupToolbar()
+    fun onToolbarLoadDataSuccess()
+    fun onToolbarLoadDataError()
+    fun onLoadStatementListSuccess()
+    fun onLoadStatementListError()
+}
