+++ BankApp/app/src/main/java/br/com/rphmelo/bankapp/currency/domain/models/StatementModel.kt
@@ -1,0 +1,6 @@
+package br.com.rphmelo.bankapp.currency.domain.models
+
+data class StatementModel( val title: String, val description: String,
+                           val date: String,  val value: Double) {
+
+}
