+++ BankApp/app/src/main/java/br/com/rphmelo/bankapp/currency/presentation/adapters/StatementListAdapter.kt
@@ -1,0 +1,25 @@
+package br.com.rphmelo.bankapp.currency.presentation.adapters
+
+import android.content.Context
+import android.support.v7.widget.RecyclerView
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import br.com.rphmelo.bankapp.R
+import br.com.rphmelo.bankapp.currency.domain.models.StatementModel
+
+class StatementListAdapter(private val list: List<StatementModel>,
+                           private val context: Context) : RecyclerView.Adapter<ViewHolder>(){
+
+    override fun onCreateViewHolder(parent: ViewGroup, p1: Int): ViewHolder {
+        val view = LayoutInflater.from(context).inflate(R.layout.item_statement_list, parent, false)
+        return ViewHolder(view)
+    }
+
+    override fun getItemCount(): Int {
+        return list.size
+    }
+
+    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
+
+    }
+}
