+++ BankApp/app/src/main/java/br/com/rphmelo/bankapp/login/domain/models/LoginError.kt
@@ -1,0 +1,3 @@
+package br.com.rphmelo.bankapp.login.domain.models
+
+data class LoginError(val code: Int, val message: String)
