+++ BankApp/app/src/main/java/br/com/rphmelo/bankapp/login/domain/models/LoginRequest.kt
@@ -1,0 +1,3 @@
+package br.com.rphmelo.bankapp.login.domain.models
+
+data class LoginRequest(val user: String, val password: String)
