+++ BankApp/app/src/main/java/br/com/rphmelo/bankapp/login/domain/models/LoginResponse.kt
@@ -1,0 +1,3 @@
+package br.com.rphmelo.bankapp.login.domain.models
+
+data class LoginResponse(val userAccount: UserAccount, val error: LoginError)
