+++ BankApp/app/src/main/java/br/com/rphmelo/bankapp/login/domain/models/LoginView.kt
@@ -1,0 +1,12 @@
+package br.com.rphmelo.bankapp.login.domain.models
+
+interface LoginView {
+    fun showProgress()
+    fun hideProgress()
+    fun setUserEmptyError()
+    fun setPasswordEmptyError()
+    fun setUserInvalidError()
+    fun setPasswordInvalidError()
+    fun onLoginSuccess(loginResponse: LoginResponse)
+    fun onLoginError()
+}
