+++ BankApp/app/src/main/java/br/com/rphmelo/bankapp/login/domain/models/OnLoginFinishedListener.kt
@@ -1,0 +1,10 @@
+package br.com.rphmelo.bankapp.login.domain.models
+
+interface OnLoginFinishedListener {
+    fun onUserEmptyError()
+    fun onPasswordEmptyError()
+    fun onUserInvalidError()
+    fun onPasswordInvalidError()
+    fun onLoginSuccess(loginResponse: LoginResponse)
+    fun onLoginError()
+}
