--- BankApp/app/src/main/java/br/com/rphmelo/bankapp/models/LoginError.kt
@@ -1,3 +1,0 @@
-package br.com.rphmelo.bankapp.models
-
-data class LoginError(val code: Int, val message: String)
