--- BankApp/app/src/main/java/br/com/rphmelo/bankapp/models/LoginRequest.kt
@@ -1,3 +1,0 @@
-package br.com.rphmelo.bankapp.models
-
-data class LoginRequest(val user: String, val password: String)
