--- BankApp/app/src/main/java/br/com/rphmelo/bankapp/models/LoginResponse.kt
@@ -1,3 +1,0 @@
-package br.com.rphmelo.bankapp.models
-
-data class LoginResponse(val userAccount: UserAccount, val error: LoginError)
