--- BankApp/app/src/main/java/br/com/rphmelo/bankapp/models/UserAccount.kt
@@ -1,4 +1,0 @@
-package br.com.rphmelo.bankapp.models
-
-data class UserAccount(val userId: Int, val name: String, val bankAccount: String,
-                       val agency: String, val balance: Double)
