--- BankApp/app/src/main/java/br/com/rphmelo/bankapp/presentation/currency/adapters/ViewHolder.kt
@@ -1,21 +1,0 @@
-package br.com.rphmelo.bankapp.presentation.currency.adapters
-
-import android.support.v7.widget.RecyclerView
-import android.view.View
-import br.com.rphmelo.bankapp.presentation.currency.models.StatementModel
-import kotlinx.android.synthetic.main.item_statement_list.view.*
-
-class ViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-
-    fun bindView(itemList: StatementModel) {
-        val title = itemView.lblItemTitle
-        val description = itemView.lblDescription
-        val date = itemView.lblDate
-        val value = itemView.lblValue
-
-        title?.text = itemList.title
-        description?.text = itemList.description
-        date?.text = itemList.date
-        value?.text = itemList.value.toString()
-    }
-}
