--- BankApp/app/src/main/java/br/com/rphmelo/bankapp/presentation/currency/models/CurrencyView.kt
@@ -1,11 +1,0 @@
-package br.com.rphmelo.bankapp.presentation.currency.models
-
-interface CurrencyView {
-    fun showProgress()
-    fun hideProgress()
-    fun setupToolbar()
-    fun setToolbarData()
-    fun setToolbarError()
-    fun setStatementListData()
-    fun setStatementListError()
-}
