--- BankApp/app/src/main/java/br/com/rphmelo/bankapp/presentation/currency/models/OnCurrencyLoadDataListener.kt
@@ -1,9 +1,0 @@
-package br.com.rphmelo.bankapp.presentation.currency.models
-
-interface OnCurrencyLoadDataListener {
-    fun onSetupToolbar()
-    fun onToolbarLoadDataSuccess()
-    fun onToolbarLoadDataError()
-    fun onLoadStatementListSuccess()
-    fun onLoadStatementListError()
-}
