--- BankApp/app/src/main/java/br/com/rphmelo/bankapp/utils/Variables.kt
@@ -1,6 +1,0 @@
-package br.com.rphmelo.bankapp.utils
-
-object Variables {
-    val ROOT_URL = "https://bank-app-test.herokuapp.com/api/"
-
-}
