+++ BankApp/app/src/main/res/layout/activity_login.xml
@@ -8,7 +8,7 @@
     <android.support.constraint.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.login.LoginActivity"
+        tools:context=".login.presentation.LoginActivity"
         >
 
         <ImageView
